refactoring + removing binaryen `gufa` flag for now
