+++ src/commonMain/kotlin/macroBenchmarks/coroutinesSlowBenchmarks/SharedFlowBaseline.kt
@@ -1,20 +1,24 @@
-package macroBenchmarks.coroutinesFastBenchmarks
+package macroBenchmarks.coroutinesSlowBenchmarks
 
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.launch
-import macroBenchmarks.coroutines.ParametrizedDispatcherBaseSlow
+import macroBenchmarks.coroutines.ParametrizedDispatcherBase
 import kotlin.coroutines.Continuation
 import kotlin.coroutines.startCoroutine
+import kotlinx.benchmark.*
 
-// Stresses out 'synchronized' codepath in MutableSharedFlow
-open class SharedFlowBaseline : ParametrizedDispatcherBaseSlow() {
+/* Adapted benchmark from kotlinx.coroutines
+ * https://github.com/Kotlin/kotlinx.coroutines/blob/master/kotlinx-coroutines-core/benchmarks/main/kotlin/SharedFlowBaseline.kt
+ * Stresses out 'synchronized' codepath in MutableSharedFlow
+ */
+@State(Scope.Benchmark)
+open class SharedFlowBaseline : ParametrizedDispatcherBase() {
 
     private var size: Int = 10_000
 
-    override fun verifyResult(result: Any) = (result is Int) && result == size * (size - 1) / 2
-
-    override fun benchmark(): Any {
+    @Benchmark
+    fun baseline() {
         var sum = 0
         var done = false
         suspend {
@@ -27,6 +31,6 @@ open class SharedFlowBaseline : ParametrizedDispatcherBaseSlow() {
         }.startCoroutine(Continuation(coroutineContext) { it.getOrThrow() })
         coroutineContext.drain()
         check(done) { "benchmark did not complete" }
-        return sum
+        check(sum == size * (size - 1) / 2) { "benchmark did not complete $sum" }
     }
 }
